pyproject.toml
[tool.poetry]
name = "bluecat-csv-importer"
version = "0.3.0"
description = "Production-grade CSV-based bulk importer for BlueCat Address Manager"
authors = ["Your Name <you@example.com>"]
readme = "docs/README.md"
packages = [{include = "importer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = ">=0.15.0"}
rich = "^13.7.0"
httpx = "^0.27.0"
pydantic = "^2.7.0"
tenacity = "^8.2.3"
diskcache = "^5.6.3"
structlog = "^24.1.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
respx = "^0.21.0"
hypothesis = "^6.100.0"
black = "^24.3.0"
ruff = "^0.3.0"
mypy = "^1.9.0"
types-pyyaml = "^6.0.12"

[tool.poetry.scripts]
bluecat-import = "importer.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # function calls in argument defaults (standard Typer pattern)
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
